+++ src/main/js/components/Application.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import HelloMessage from "./HelloMessage";
-import HomeTimelineUI from "./HomeTimelineUI";
-import UserTimelineUI from "./UserTimelineUI";
+import HomeTimelineUI from "./timeline/HomeTimelineUI";
+import UserTimelineUI from "./timeline/UserTimelineUI";
 import "../../css/components/Application.scss";
 
 export default class Application extends React.Component {
