+++ src/main/js/components/timeline/UserTimelineUI.js
@@ -1,0 +1,18 @@
+import React, { Component } from "react";
+import TimelineUI from "./TimelineUI";
+import { fetchUserTimeline } from "../../twitter-api";
+
+export default class UserTimelineUI extends React.Component {
+
+	static get buttonName() {
+		return "Update User Timeline";
+	}
+
+	render() {
+		return (
+			<React.Fragment>
+				<TimelineUI apiCall={fetchUserTimeline} buttonName={UserTimelineUI.buttonName}/>
+			</React.Fragment>
+		);
+	}
+}
