+++ src/test/js/components/timeline/TimelineUI.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { shallow } from "enzyme";
-import TimelineUI from "../../../main/js/components/TimelineUI";
-import { expectOne } from "../test-util";
+import TimelineUI from "../../../../main/js/components/timeline/TimelineUI";
+import { expectOne } from "../../test-util";
 
 describe("TimelineUI", () => {
